+++ Tunny/Solver/Optuna/Optuna.cs
@@ -12,18 +12,18 @@ using Tunny.Settings;
 using Tunny.UI;
 using Tunny.Util;
 
-namespace Tunny.Solver
+namespace Tunny.Solver.Optuna
 {
     public class Optuna
     {
         public double[] XOpt { get; private set; }
-        private double[] FxOpt { get; set; }
-
         private readonly string _componentFolder;
+        private readonly TunnySettings _settings;
 
-        public Optuna(string componentFolder)
+        public Optuna(string componentFolder, TunnySettings settings)
         {
             _componentFolder = componentFolder;
+            _settings = settings;
             string envPath = PythonInstaller.GetEmbeddedPythonPath() + @"\python310.dll";
             Environment.SetEnvironmentVariable("PYTHONNET_PYDLL", envPath, EnvironmentVariableTarget.Process);
         }
@@ -31,23 +31,9 @@ namespace Tunny.Solver
         public bool RunSolver(
             List<Variable> variables,
             IEnumerable<IGH_Param> objectives,
-            Func<IList<decimal>, int, EvaluatedGHResult> evaluate,
-            TunnySettings settings)
+            Func<IList<decimal>, int, EvaluatedGHResult> evaluate)
         {
-            int dVar = variables.Count;
-            double[] lb = new double[dVar];
-            double[] ub = new double[dVar];
-            bool[] isInteger = new bool[dVar];
-            string[] varNickName = new string[dVar];
             string[] objNickName = objectives.Select(x => x.NickName).ToArray();
-
-            for (int i = 0; i < dVar; i++)
-            {
-                lb[i] = Convert.ToDouble(variables[i].LowerBond);
-                ub[i] = Convert.ToDouble(variables[i].UpperBond);
-                isInteger[i] = variables[i].IsInteger;
-                varNickName[i] = variables[i].NickName;
-            }
 
             EvaluatedGHResult Eval(double[] x, int progress)
             {
@@ -57,29 +43,48 @@ namespace Tunny.Solver
 
             try
             {
-                var tpe = new OptunaAlgorithm(lb, ub, varNickName, objNickName, settings, Eval);
-                tpe.Solve();
-                XOpt = tpe.GetXOptimum();
-                FxOpt = tpe.GetFxOptimum();
+                var optimize = new Algorithm(variables, objNickName, _settings, Eval);
+                optimize.Solve();
+                XOpt = optimize.GetXOptimum();
 
-                TunnyMessageBox.Show("Solver completed successfully.", "Tunny");
-
+                ShowEndMessages(optimize);
                 return true;
             }
             catch (Exception e)
             {
-                TunnyMessageBox.Show(
-                    "Tunny runtime error:\n" +
-                    "Please send below message (& gh file if possible) to Tunny support.\n\n" +
-                    "\" " + e.Message + " \"", "Tunny",
-                    MessageBoxButtons.OK, MessageBoxIcon.Error);
+                ShowErrorMessages(e);
                 return false;
             }
         }
 
+        private static void ShowEndMessages(Algorithm optimize)
+        {
+            switch (optimize.EndState)
+            {
+                case EndState.Timeout:
+                    TunnyMessageBox.Show("Solver completed successfully.\nThe specified time has elapsed.", "Tunny");
+                    break;
+                case EndState.AllTrialFinish:
+                    TunnyMessageBox.Show("Solver completed successfully.\nThe specified number of trials has been completed.", "Tunny");
+                    break;
+                default:
+                    TunnyMessageBox.Show("Solver error.", "Tunny");
+                    break;
+            }
+        }
+
+        private static void ShowErrorMessages(Exception e)
+        {
+            TunnyMessageBox.Show(
+                "Tunny runtime error:\n" +
+                "Please send below message (& gh file if possible) to Tunny support.\n\n" +
+                "\" " + e.Message + " \"", "Tunny",
+                MessageBoxButtons.OK, MessageBoxIcon.Error);
+        }
+
         public void ShowSelectedTypePlot(string visualize, string studyName)
         {
-            string storage = "sqlite:///" + _componentFolder + "/Tunny_Opt_Result.db";
+            string storage = "sqlite:///" + _settings.Storage;
             PythonEngine.Initialize();
             using (Py.GIL())
             {
@@ -98,45 +103,7 @@ namespace Tunny.Solver
                 string[] nickNames = ((string)study.user_attrs["objective_names"]).Split(',');
                 try
                 {
-                    dynamic vis;
-                    switch (visualize)
-                    {
-                        case "contour":
-                            vis = optuna.visualization.plot_contour(study, target_name: nickNames[0]);
-                            vis.show();
-                            break;
-                        case "EDF":
-                            vis = optuna.visualization.plot_edf(study, target_name: nickNames[0]);
-                            vis.show();
-                            break;
-                        case "intermediate values":
-                            vis = optuna.visualization.plot_intermediate_values(study);
-                            vis.show();
-                            break;
-                        case "optimization history":
-                            vis = optuna.visualization.plot_optimization_history(study, target_name: nickNames[0]);
-                            vis.show();
-                            break;
-                        case "parallel coordinate":
-                            vis = optuna.visualization.plot_parallel_coordinate(study, target_name: nickNames[0]);
-                            vis.show();
-                            break;
-                        case "param importances":
-                            vis = optuna.visualization.plot_param_importances(study, target_name: nickNames[0]);
-                            vis.show();
-                            break;
-                        case "pareto front":
-                            vis = optuna.visualization.plot_pareto_front(study, target_names: nickNames);
-                            vis.show();
-                            break;
-                        case "slice":
-                            vis = optuna.visualization.plot_slice(study, target_name: nickNames[0]);
-                            vis.show();
-                            break;
-                        default:
-                            TunnyMessageBox.Show("This visualization type is not supported in this study case.", "Tunny");
-                            break;
-                    }
+                    ShowPlot(optuna, visualize, study, nickNames);
                 }
                 catch (Exception)
                 {
@@ -146,9 +113,45 @@ namespace Tunny.Solver
             PythonEngine.Shutdown();
         }
 
+        private static void ShowPlot(dynamic optuna, string visualize, dynamic study, string[] nickNames)
+        {
+            dynamic vis;
+            switch (visualize)
+            {
+                case "contour":
+                    vis = optuna.visualization.plot_contour(study, target_name: nickNames[0]);
+                    break;
+                case "EDF":
+                    vis = optuna.visualization.plot_edf(study, target_name: nickNames[0]);
+                    break;
+                case "intermediate values":
+                    vis = optuna.visualization.plot_intermediate_values(study);
+                    break;
+                case "optimization history":
+                    vis = optuna.visualization.plot_optimization_history(study, target_name: nickNames[0]);
+                    break;
+                case "parallel coordinate":
+                    vis = optuna.visualization.plot_parallel_coordinate(study, target_name: nickNames[0]);
+                    break;
+                case "param importances":
+                    vis = optuna.visualization.plot_param_importances(study, target_name: nickNames[0]);
+                    break;
+                case "pareto front":
+                    vis = optuna.visualization.plot_pareto_front(study, target_names: nickNames);
+                    break;
+                case "slice":
+                    vis = optuna.visualization.plot_slice(study, target_name: nickNames[0]);
+                    break;
+                default:
+                    TunnyMessageBox.Show("This visualization type is not supported in this study case.", "Tunny");
+                    return;
+            }
+            vis.show();
+        }
+
         public ModelResult[] GetModelResult(int[] resultNum, string studyName)
         {
-            string storage = "sqlite:///" + _componentFolder + "/Tunny_Opt_Result.db";
+            string storage = "sqlite:///" + _settings.Storage;
             var modelResult = new List<ModelResult>();
             PythonEngine.Initialize();
             using (Py.GIL())
@@ -166,34 +169,39 @@ namespace Tunny.Solver
                     return modelResult.ToArray();
                 }
 
-                if (resultNum[0] == -1)
-                {
-                    var bestTrials = (dynamic[])study.best_trials;
-                    foreach (dynamic trial in bestTrials)
-                    {
-                        ParseTrial(modelResult, trial);
-                    }
-                }
-                else if (resultNum[0] == -10)
-                {
-                    var trials = (dynamic[])study.trials;
-                    foreach (dynamic trial in trials)
-                    {
-                        ParseTrial(modelResult, trial);
-                    }
-                }
-                else
-                {
-                    foreach (int res in resultNum)
-                    {
-                        dynamic trial = study.trials[res];
-                        ParseTrial(modelResult, trial);
-                    }
-                }
+                SetTrialsToModelResult(resultNum, modelResult, study);
             }
             PythonEngine.Shutdown();
 
             return modelResult.ToArray();
+        }
+
+        private void SetTrialsToModelResult(int[] resultNum, List<ModelResult> modelResult, dynamic study)
+        {
+            if (resultNum[0] == -1)
+            {
+                var bestTrials = (dynamic[])study.best_trials;
+                foreach (dynamic trial in bestTrials)
+                {
+                    ParseTrial(modelResult, trial);
+                }
+            }
+            else if (resultNum[0] == -10)
+            {
+                var trials = (dynamic[])study.trials;
+                foreach (dynamic trial in trials)
+                {
+                    ParseTrial(modelResult, trial);
+                }
+            }
+            else
+            {
+                foreach (int res in resultNum)
+                {
+                    dynamic trial = study.trials[res];
+                    ParseTrial(modelResult, trial);
+                }
+            }
         }
 
         private static void ParseTrial(ICollection<ModelResult> modelResult, dynamic trial)