+++ Tunny/UI/OptimizationWindow.cs
@@ -83,6 +83,7 @@ namespace Tunny.UI
         {
             samplerComboBox.SelectedIndex = _settings.Optimize.SelectSampler;
             nTrialNumUpDown.Value = _settings.Optimize.NumberOfTrials;
+            timeoutNumUpDown.Value = (decimal)_settings.Optimize.Timeout;
             loadIfExistsCheckBox.Checked = _settings.Optimize.LoadExistStudy;
             studyNameTextBox.Text = _settings.StudyName;
             outputModelNumTextBox.Text = _settings.Result.OutputNumberString;
@@ -127,6 +128,7 @@ namespace Tunny.UI
         {
             _settings.Optimize.SelectSampler = samplerComboBox.SelectedIndex;
             _settings.Optimize.NumberOfTrials = (int)nTrialNumUpDown.Value;
+            _settings.Optimize.Timeout = (double)timeoutNumUpDown.Value;
             _settings.Optimize.LoadExistStudy = loadIfExistsCheckBox.Checked;
             _settings.StudyName = studyNameTextBox.Text;
             _settings.Result.OutputNumberString = outputModelNumTextBox.Text;